+++ GrailDiary/BookSearchViewController.swift
@@ -4,30 +4,25 @@ import Combine
 import Logging
 import UIKit
 
-public struct Book: Hashable {
-  var id: String
-  var title: String
-  var authors: [String]
-  var publishedDate: String?
+private struct ViewModel: Hashable, Identifiable {
+  var id = UUID()
+  var book: Book
   var coverImage: UIImage?
   var coverImageURL: URL?
 
-  var markdownTitle: String {
-    var title = "_\(self.title)_"
-    if !authors.isEmpty {
-      let authors = self.authors.joined(separator: ", ")
-      title += ": \(authors)"
+  init?(_ item: BookSearchViewController.GoogleBooksItem) {
+    guard let book = Book(item) else { return nil }
+    self.book = book
+    self.coverImage = nil
+    if let urlString = item.volumeInfo.imageLinks?.smallThumbnail ?? item.volumeInfo.imageLinks?.thumbnail {
+      self.coverImageURL = URL(string: urlString)
     }
-    if let publishedDate = publishedDate {
-      title += " (\(publishedDate.prefix(4)))"
-    }
-    return title
   }
 }
 
 public protocol BookSearchViewControllerDelegate: AnyObject {
   /// The person selected a book.
-  func bookSearchViewController(_ viewController: BookSearchViewController, didSelect book: Book)
+  func bookSearchViewController(_ viewController: BookSearchViewController, didSelect book: Book, coverImage: UIImage?)
 
   /// The person canceled without selecting a book.
   /// (Note this is not guaranteed to be called with the pull-down presentation style)
@@ -72,19 +67,18 @@ public final class BookSearchViewController: UIViewController {
     return collectionView
   }()
 
-  private lazy var dataSource: UICollectionViewDiffableDataSource<Int, Book> = {
-    let bookRegistration = UICollectionView.CellRegistration<UICollectionViewListCell, Book> { cell, _, book in
+  private lazy var dataSource: UICollectionViewDiffableDataSource<Int, ViewModel> = {
+    let bookRegistration = UICollectionView.CellRegistration<UICollectionViewListCell, ViewModel> { cell, _, viewModel in
       var configuration = cell.defaultContentConfiguration()
-      if let publishedDate = book.publishedDate {
-        let year = publishedDate.prefix(4)
-        configuration.text = "\(book.title) (\(year))"
+      if let year = viewModel.book.yearPublished {
+        configuration.text = "\(viewModel.book.title) (\(year))"
       } else {
-        configuration.text = book.title
+        configuration.text = viewModel.book.title
       }
       configuration.textProperties.color = .label
-      configuration.secondaryText = book.authors.joined(separator: ", ")
+      configuration.secondaryText = viewModel.book.authors.joined(separator: ", ")
       configuration.secondaryTextProperties.color = .secondaryLabel
-      configuration.image = book.coverImage
+      configuration.image = viewModel.coverImage
       var backgroundConfiguration = UIBackgroundConfiguration.listPlainCell()
       backgroundConfiguration.backgroundColor = .grailBackground
 
@@ -92,7 +86,7 @@ public final class BookSearchViewController: UIViewController {
       cell.backgroundConfiguration = backgroundConfiguration
     }
 
-    return UICollectionViewDiffableDataSource<Int, Book>(collectionView: collectionView) { collectionView, indexPath, book in
+    return UICollectionViewDiffableDataSource<Int, ViewModel>(collectionView: collectionView) { collectionView, indexPath, book in
       collectionView.dequeueConfiguredReusableCell(using: bookRegistration, for: indexPath, item: book)
     }
   }()
@@ -100,14 +94,14 @@ public final class BookSearchViewController: UIViewController {
   private let currentSearchTerm = CurrentValueSubject<String, Never>("")
   private var currentSearch: AnyCancellable?
   private let imageCache = ImageCache()
-  private var viewModels = [Book]()
+  private var viewModels = [ViewModel]()
   private lazy var decoder: JSONDecoder = {
     let decoder = JSONDecoder()
     decoder.dateDecodingStrategy = .iso8601
     return decoder
   }()
 
-  private func updateViewModels(_ viewModels: [Book]) {
+  private func updateViewModels(_ viewModels: [ViewModel]) {
     assert(Thread.isMainThread)
     isBatchUpdating = true
     self.viewModels = viewModels
@@ -123,7 +117,7 @@ public final class BookSearchViewController: UIViewController {
     updateSnapshot()
   }
 
-  private func setViewModelImage(_ image: UIImage, key: String) {
+  private func setViewModelImage(_ image: UIImage, key: UUID) {
     assert(Thread.isMainThread)
     guard let index = viewModels.firstIndex(where: { $0.id == key }) else { return }
     viewModels[index].coverImage = image
@@ -135,7 +129,7 @@ public final class BookSearchViewController: UIViewController {
   private func updateSnapshot() {
     assert(Thread.isMainThread)
     guard !isBatchUpdating else { return }
-    var snapshot = NSDiffableDataSourceSnapshot<Int, Book>()
+    var snapshot = NSDiffableDataSourceSnapshot<Int, ViewModel>()
     snapshot.appendSections([0])
     snapshot.appendItems(viewModels.removingDuplicates())
     dataSource.apply(snapshot, animatingDifferences: true)
@@ -190,7 +184,7 @@ public final class BookSearchViewController: UIViewController {
         }
       } receiveValue: { [weak self] data in
         DispatchQueue.main.async {
-          self?.updateViewModels(data.items.map { Book($0) })
+          self?.updateViewModels(data.items.compactMap { ViewModel($0) })
         }
       }
   }
@@ -202,8 +196,8 @@ public final class BookSearchViewController: UIViewController {
 
 extension BookSearchViewController: UICollectionViewDelegate {
   public func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-    guard let book = dataSource.itemIdentifier(for: indexPath) else { return }
-    delegate?.bookSearchViewController(self, didSelect: book)
+    guard let viewModel = dataSource.itemIdentifier(for: indexPath) else { return }
+    delegate?.bookSearchViewController(self, didSelect: viewModel.book, coverImage: viewModel.coverImage)
   }
 }
 
@@ -220,12 +214,27 @@ private extension BookSearchViewController {
     var volumeInfo: VolumeInfo
   }
 
+  enum IndustryIdentifierType: String, Codable {
+    case isbn10 = "ISBN_10"
+    case isbn13 = "ISBN_13"
+    case issn = "ISSN"
+    case other = "OTHER"
+  }
+
+  struct IndustryIdentifier: Codable {
+    var type: IndustryIdentifierType
+    var identifier: String
+  }
+
   struct VolumeInfo: Codable {
-    var title: String
+    var title: String?
     var subtitle: String?
     var authors: [String]?
     var publishedDate: String?
     var imageLinks: ImageLink?
+    var industryIdentifiers: [IndustryIdentifier]?
+    var pageCount: Int?
+    var publisher: String?
   }
 
   struct ImageLink: Codable {
@@ -235,14 +244,29 @@ private extension BookSearchViewController {
 }
 
 private extension Book {
-  init(_ item: BookSearchViewController.GoogleBooksItem) {
-    self.id = item.id
-    self.title = item.volumeInfo.title
+  /// Construct a Book model from a Google Books search result.
+  init?(_ item: BookSearchViewController.GoogleBooksItem) {
+    guard let title = item.volumeInfo.title else {
+      Logger.shared.info("Item without title: \(item)")
+      return nil
+    }
+    self.title = title
     self.authors = item.volumeInfo.authors ?? []
-    self.publishedDate = item.volumeInfo.publishedDate
-    self.coverImage = nil
-    if let urlString = item.volumeInfo.imageLinks?.smallThumbnail ?? item.volumeInfo.imageLinks?.thumbnail {
-      self.coverImageURL = URL(string: urlString)
+    self.numberOfPages = item.volumeInfo.pageCount
+    self.publisher = item.volumeInfo.publisher
+    if let datePrefix = item.volumeInfo.publishedDate?.prefix(4) {
+      self.yearPublished = Int(datePrefix)
+    }
+    for identifier in item.volumeInfo.industryIdentifiers ?? [] {
+      switch identifier.type {
+      case .isbn10:
+        self.isbn = identifier.identifier
+      case .isbn13:
+        self.isbn13 = identifier.identifier
+      case .issn, .other:
+        // ignore
+        break
+      }
     }
   }
 }