+++ GrailDiary/Note+Database.swift
@@ -4,11 +4,6 @@ import Foundation
 import GRDB
 import Logging
 
-private enum ApplicationMimeType: String {
-  /// Private MIME type for URLs.
-  case url = "text/vnd.grail.url"
-}
-
 public extension Note {
   // TODO: Should this all be done in a single massive join? Instead there are lots of database round-trips.
   /// Loads a note from the database.
@@ -84,7 +79,9 @@ public extension Note {
       .fetchAll(db)
       .map { $0.key }
       .asSet()
-    let obsoleteImages = onDiskImageKeys.subtracting(referencedImageKeys)
+    var obsoleteImages = onDiskImageKeys.subtracting(referencedImageKeys)
+    // An image with the special key "coverImage" is always considered referenced; never remove it.
+    obsoleteImages.remove(Note.coverImageKey)
     try obsoleteImages.forEach { imageKey in
       try BinaryContentRecord.deleteOne(db, key: ["noteId": identifier, "key": imageKey])
     }
@@ -152,17 +149,28 @@ public extension Note {
 
   private func saveReference(noteIdentifier: Note.Identifier, database: Database) throws {
     guard let reference = reference else { return }
+    let record: ContentRecord
     switch reference {
     case .webPage(let url):
-      let record = ContentRecord(
+      record = ContentRecord(
         text: url.absoluteString,
         noteId: noteIdentifier,
         key: ContentRole.reference.rawValue,
         role: ContentRole.reference.rawValue,
         mimeType: ApplicationMimeType.url.rawValue
       )
-      try record.save(database)
+    case .book(let book):
+      let bookData = try JSONEncoder().encode(book)
+      let bookString = String(data: bookData, encoding: .utf8)!
+      record = ContentRecord(
+        text: bookString,
+        noteId: noteIdentifier,
+        key: ContentRole.reference.rawValue,
+        role: ContentRole.reference.rawValue,
+        mimeType: ApplicationMimeType.book.rawValue
+      )
     }
+    try record.save(database)
   }
 }
 
@@ -187,6 +195,10 @@ private extension Database {
         Logger.shared.error("Could not turn string into URL: \(record.text)")
         return nil
       }
+    case .book:
+      let bookData = record.text.data(using: .utf8)!
+      let book = try JSONDecoder().decode(Book.self, from: bookData)
+      return .book(book)
     }
   }
 }