+++ GrailDiary/NotebookStructureViewController.swift
@@ -142,6 +142,11 @@ final class NotebookStructureViewController: UIViewController {
   public weak var delegate: NotebookStructureViewControllerDelegate?
   private let database: NoteDatabase
   private var notebookSubscription: AnyCancellable?
+  private var progressView: UIProgressView? {
+    didSet {
+      configureToolbar()
+    }
+  }
 
   private lazy var collectionView: UICollectionView = {
     let layout = UICollectionViewCompositionalLayout { sectionIndex, layoutEnvironment in
@@ -419,20 +424,59 @@ extension NotebookStructureViewController: UICollectionViewDelegate {
           Logger.shared.error("Error moving notes tagged \(hashtag) to inbox: \(error)")
         }
       }
-      return UIMenu(title: "", children: [rename, moveToInboxAction])
+      let moveToTrashAction = UIAction(title: "Delete", image: UIImage(systemName: "trash")) { _ in
+        do {
+          try database.moveNotesTaggedWithHashtag(hashtag, to: PredefinedFolder.recentlyDeleted.rawValue)
+        } catch {
+          Logger.shared.error("Error moving notes tagged \(hashtag) to inbox: \(error)")
+        }
+      }
+      return UIMenu(title: "", children: [rename, moveToInboxAction, moveToTrashAction])
     }
   }
 }
 
 // MARK: - Private
 
+private extension UIBarButtonItem {
+  static func documentBrowser() -> UIBarButtonItem {
+    let button = UIBarButtonItem(title: "Open", style: .plain, target: nil, action: #selector(AppCommands.openNewFile))
+    button.accessibilityIdentifier = "open-files"
+    return button
+  }
+}
+
 private extension NotebookStructureViewController {
   func configureToolbar() {
-    var toolbarItems = [AppCommandsButtonItems.documentBrowser(), UIBarButtonItem.flexibleSpace()]
+    var toolbarItems: [UIBarButtonItem] = [documentActionMenu(), .flexibleSpace()]
+    if let progressView = progressView {
+      toolbarItems.append(UIBarButtonItem(customView: progressView))
+      toolbarItems.append(.flexibleSpace())
+    }
     if splitViewController?.isCollapsed ?? false, let newNoteButton = notebookViewController?.makeNewNoteButtonItem() {
       toolbarItems.append(newNoteButton)
     }
     self.toolbarItems = toolbarItems
+  }
+
+  private func documentActionMenu() -> UIBarButtonItem {
+    let openCommand = UICommand(title: "Open", image: UIImage(systemName: "doc"), action: #selector(AppCommands.openNewFile))
+    let importLibraryThing = UIAction(title: "Import LibraryThing", image: UIImage(systemName: "arrow.down.doc")) { [weak self] _ in
+      guard let self = self else { return }
+      Logger.shared.info("Importing from LibraryThing")
+      let bookImporterViewController = BookImporterViewController(database: self.database)
+      bookImporterViewController.delegate = self
+      self.present(bookImporterViewController, animated: true)
+    }
+    return UIBarButtonItem(
+      image: UIImage(systemName: "ellipsis.circle"),
+      menu: UIMenu(
+        children: [
+          openCommand,
+          importLibraryThing,
+        ]
+      )
+    )
   }
 
   func updateSnapshot() {
@@ -488,5 +532,21 @@ private extension NotebookStructureViewController {
       snapshot.append([stringToItem[hashtag]!], to: parent ?? root)
     }
     return snapshot
+  }
+}
+
+extension NotebookStructureViewController: BookImporterViewControllerDelegate {
+  func bookImporter(_ bookImporter: BookImporterViewController, didStartImporting count: Int) {
+    let progressView = UIProgressView(progressViewStyle: .bar)
+    progressView.progress = 0
+    self.progressView = progressView
+  }
+
+  func bookImporter(_ bookImporter: BookImporterViewController, didProcess partialCount: Int, of totalCount: Int) {
+    progressView?.progress = Float(partialCount) / Float(totalCount)
+  }
+
+  func bookImporterDidFinishImporting(_ bookImporter: BookImporterViewController) {
+    progressView = nil
   }
 }